Stabilize admin flow and remove debug-side effects
